+++ src/Api/Graph/Queries/CategoryQuery.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace Expressionengine\Coilpack\Api\Graph\Queries;
+
+use Expressionengine\Coilpack\Models\Category\Category;
+use GraphQL\Type\Definition\Type;
+use Rebing\GraphQL\Support\Facades\GraphQL;
+use Rebing\GraphQL\Support\Query;
+
+class CategoryQuery extends Query
+{
+    protected $attributes = [
+        'name' => 'category',
+    ];
+
+    public function type(): Type
+    {
+        return GraphQL::type('Category');
+    }
+
+    public function args(): array
+    {
+        return [
+            'category_id' => [
+                'name' => 'category_id',
+                'type' => Type::int(),
+                'rules' => ['required'],
+            ],
+        ];
+    }
+
+    public function resolve($root, $args)
+    {
+        return Category::findOrFail($args['category_id']);
+    }
+}